@@ -1,8 +1,8 @@
 [package]
 name = "firefly_cli"
 version = "0.15.1"
-rust-version = "1.82.0"
-edition = "2021"
+rust-version = "1.91.0"
+edition = "2024"
 authors = ["Firefly Zero team"]
 description = "CLI tool for working with Firefly Zero"
 homepage = "https://fireflyzero.com/"
