@@ -2,18 +2,18 @@ use crate::args::ImportArgs;
 use crate::crypto::hash_dir;
 use crate::file_names::{HASH, KEY, META, SIG, STATS};
 use crate::vfs::init_vfs;
-use anyhow::{bail, Context, Result};
+use anyhow::{Context, Result, bail};
 use chrono::Datelike;
 use data_encoding::HEXLOWER;
 use firefly_types::{Encode, Meta};
+use rsa::RsaPublicKey;
 use rsa::pkcs1::DecodeRsaPublicKey;
 use rsa::pkcs1v15::{Signature, VerifyingKey};
 use rsa::signature::hazmat::PrehashVerifier;
-use rsa::RsaPublicKey;
 use serde::Deserialize;
 use sha2::Sha256;
 use std::env::temp_dir;
-use std::fs::{self, create_dir_all, File};
+use std::fs::{self, File, create_dir_all};
 use std::io::Read;
 use std::path::{Path, PathBuf};
 use zip::ZipArchive;
