@@ -1,5 +1,5 @@
 use crate::args::RuntimeArgs;
-use crate::net::{connect, is_timeout, Stream};
+use crate::net::{Stream, connect, is_timeout};
 use anyhow::{Context, Result};
 use crossterm::{cursor, event, execute, style, terminal};
 use firefly_types::serial;
