@@ -1,5 +1,5 @@
 use crate::args::ShotsDownloadArgs;
-use anyhow::{bail, Context, Result};
+use anyhow::{Context, Result, bail};
 use std::io::Write;
 use std::path::{Path, PathBuf};
 
