@@ -1,4 +1,4 @@
-use anyhow::{bail, Context};
+use anyhow::{Context, bail};
 use std::io::Write;
 use std::path::Path;
 use std::process::Command;
